refactor(linter): `message_to_message_with_position` helper function (#13140)
Is needed later for `TsGoLintState`
