+++ crates/oxc_linter/src/fixer/mod.rs
@@ -11,6 +11,8 @@ use crate::service::offset_to_position::SpanPositionMessage;
 #[cfg(feature = "language_server")]
 pub use fix::{FixWithPosition, PossibleFixesWithPosition};
 #[cfg(feature = "language_server")]
+use oxc_data_structures::rope::Rope;
+#[cfg(feature = "language_server")]
 use oxc_diagnostics::{OxcCode, Severity};
 
 mod fix;
@@ -257,6 +259,68 @@ impl From<OxcDiagnostic> for MessageWithPosition<'_> {
             url: from.url.clone(),
             fixes: PossibleFixesWithPosition::None,
         }
+    }
+}
+
+// clippy: the source field is checked and assumed to be less than 4GB, and
+// we assume that the fix offset will not exceed 2GB in either direction
+#[cfg(feature = "language_server")]
+#[expect(clippy::cast_possible_truncation)]
+pub fn message_to_message_with_position<'a>(
+    message: &Message<'a>,
+    source_text: &str,
+    rope: &Rope,
+) -> MessageWithPosition<'a> {
+    use crate::service::offset_to_position::offset_to_position;
+
+    let labels = message.error.labels.as_ref().map(|labels| {
+        labels
+            .iter()
+            .map(|labeled_span| {
+                let offset = labeled_span.offset() as u32;
+                let start_position = offset_to_position(rope, offset, source_text);
+                let end_position =
+                    offset_to_position(rope, offset + labeled_span.len() as u32, source_text);
+                let message = labeled_span.label().map(|label| Cow::Owned(label.to_string()));
+
+                SpanPositionMessage::new(start_position, end_position).with_message(message)
+            })
+            .collect::<Vec<_>>()
+    });
+
+    MessageWithPosition {
+        message: message.error.message.clone(),
+        severity: message.error.severity,
+        help: message.error.help.clone(),
+        url: message.error.url.clone(),
+        code: message.error.code.clone(),
+        labels,
+        fixes: match &message.fixes {
+            PossibleFixes::None => PossibleFixesWithPosition::None,
+            PossibleFixes::Single(fix) => {
+                PossibleFixesWithPosition::Single(fix_to_fix_with_position(fix, rope, source_text))
+            }
+            PossibleFixes::Multiple(fixes) => PossibleFixesWithPosition::Multiple(
+                fixes.iter().map(|fix| fix_to_fix_with_position(fix, rope, source_text)).collect(),
+            ),
+        },
+    }
+}
+
+#[cfg(feature = "language_server")]
+fn fix_to_fix_with_position<'a>(
+    fix: &Fix<'a>,
+    rope: &Rope,
+    source_text: &str,
+) -> FixWithPosition<'a> {
+    use crate::service::offset_to_position::offset_to_position;
+
+    let start_position = offset_to_position(rope, fix.span.start, source_text);
+    let end_position = offset_to_position(rope, fix.span.end, source_text);
+    FixWithPosition {
+        content: fix.content.clone(),
+        span: SpanPositionMessage::new(start_position, end_position)
+            .with_message(fix.message.as_ref().map(|label| Cow::Owned(label.to_string()))),
     }
 }
